requirements.txt
together
python-dotenv
Pillow
numpy
